Fix VC build error

+++ cpp/oneapi/dal/detail/parameters/system_parameters_impl.cpp
@@ -40,7 +40,8 @@ system_parameters_impl::system_parameters_impl() {
     sys_info_.insert(
         { "top_enabled_cpu_extension", std::make_any<cpu_extension>(from_daal_cpu_type(cpuid)) });
     sys_info_.insert(
-        { "max_number_of_threads", std::make_any<std::uint32_t>(env->getNumberOfThreads()) });
+        { "max_number_of_threads",
+          std::make_any<std::uint32_t>(static_cast<std::uint32_t>(env->getNumberOfThreads())) });
 }
 
 cpu_extension system_parameters_impl::get_top_enabled_cpu_extension() const {
